move Greatest and Least from conditionalExpressions.scala to arithmetic.scala
